+++ src/orquestra/sdk/_base/_driver/_ce_runtime.py
@@ -136,10 +136,6 @@ class CERuntime(RuntimeInterface):
             )
         return workflow_run_id
 
-    def get_all_workflow_runs_status(self) -> List[WorkflowRun]:
-        """Gets the status of all workflow runs."""
-        raise NotImplementedError()
-
     def get_workflow_run_status(self, workflow_run_id: WorkflowRunId) -> WorkflowRun:
         """
         Gets the status of a workflow run
@@ -166,22 +162,6 @@ class CERuntime(RuntimeInterface):
                 f"`{workflow_run_id}` "
                 "- the authorization token was rejected by the remote cluster."
             ) from e
-
-    def get_workflow_run_outputs(self, workflow_run_id: WorkflowRunId) -> Sequence[Any]:
-        """Returns the output artifacts of a workflow run
-
-        For example, for this workflow:
-
-            @sdk.workflow
-            def my_wf():
-                return [my_task(), another_task()]
-
-        this method will return an iterable that yields the results from my_task and
-        another_task().
-
-        This method blocks until the workflow is completed
-        """
-        raise NotImplementedError()
 
     def get_workflow_run_outputs_non_blocking(
         self, workflow_run_id: WorkflowRunId