+++ src/orquestra/sdk/_ray/_dag.py
@@ -783,16 +783,6 @@ class RayRuntime(RuntimeInterface):
 
         return wf_run_id
 
-    def get_all_workflow_runs_status(self) -> t.List[WorkflowRun]:
-        all_workflows = self._client.list_all()
-        wf_runs = []
-        for run_id, _ in all_workflows:
-            try:
-                wf_runs.append(self.get_workflow_run_status(run_id))
-            except exceptions.NotFoundError as e:
-                raise e
-        return wf_runs
-
     def get_workflow_run_status(self, workflow_run_id: WorkflowRunId) -> WorkflowRun:
         """
         Raises:
@@ -842,16 +832,6 @@ class RayRuntime(RuntimeInterface):
             ),
         )
 
-    def get_workflow_run_outputs(
-        self, workflow_run_id: WorkflowRunId
-    ) -> t.Sequence[t.Any]:
-        try:
-            return self._client.get_workflow_output(workflow_run_id)
-        except ValueError as e:
-            raise exceptions.NotFoundError(
-                f"Workflow run {workflow_run_id} wasn't found"
-            ) from e
-
     def get_workflow_run_outputs_non_blocking(
         self, workflow_run_id: WorkflowRunId
     ) -> t.Sequence[t.Any]:
@@ -862,7 +842,12 @@ class RayRuntime(RuntimeInterface):
                 f"{workflow_run_id} has not succeeded. {workflow_state}",
                 workflow_state,
             )
-        return self.get_workflow_run_outputs(workflow_run_id)
+        try:
+            return self._client.get_workflow_output(workflow_run_id)
+        except ValueError as e:
+            raise exceptions.NotFoundError(
+                f"Workflow run {workflow_run_id} wasn't found"
+            ) from e
 
     def get_available_outputs(
         self, workflow_run_id: WorkflowRunId